Home

NLIDatasets.jl

NLIDatasets.jl is a Julia package for working with datasets for the Natural Language Inference task.

Provides an interface to the following datasets:

SNLI

NLIDatasets.SNLI — Type.

SNLI

A corpus of 570k human-written English sentence pairs for NLI.

SNLI sentence pairs are manually labeled with labels entailment, contradiction, and neutral.

For details, see the SNLI home page or read the 2015 paper "A large annotated corpus for learning natural language inference" by Samuel R. Bowman, Gabor Angeli, Christopher Potts, and Christopher Manning.

source

MultiNLI

NLIDatasets.MultiNLI — Type.

MultiNLI

A corpus of 433k sentence pairs for NLI.

For details, see the MultiNLI home page or read the 2018 paper "A Broad-Coverage Challenge Corpus for Sentence Understanding through Inference" by Adina Williams, NIkita Nangia, and Samuel R. Bowman.

source

XNLI

NLIDatasets.XNLI — Type.

XNLI

A collection of 5,000 test and 2,500 dev pairs for the MultiNLI corpus.

For details, see the (2018 paper)[https://www.aclweb.org/anthology/papers/D/D18/D18-1269/] "XNLI: Evaluating Cross-lingual Sentence Representations."

source

HANS

NLIDatasets.HANS — Type.

HANS

HANS (Heauristic Analysis for NLI Systems) is a dataset for NLI.

For details, see the paper "Right for the Wrong Reasons: Diagnosing Syntactic Heuristics in Natural Language Inference" by by R. Tom McCoy, Ellie Pavlick, and Tal Linzen.

Consists of a set of examples for evaluation, provided with test_tsv.

source